+++ src/js/ui/tab-maps.js
@@ -11,7 +11,6 @@ const listfile = require('../casc/listfile');
 const constants = require('../constants');
 
 const WDCReader = require('../db/WDCReader');
-
 const BLPFile = require('../casc/blp');
 const WDTLoader = require('../3D/loaders/WDTLoader');
 const ADTExporter = require('../3D/exporters/ADTExporter');
@@ -184,7 +183,11 @@ const exportSelectedMapWMO = async () => {
 		const wmo = new WMOExporter(data, fileDataID);
 
 		wmo.setDoodadSetMask({ [placement.doodadSetIndex]: { checked: true } });
-		await wmo.exportAsOBJ(exportPath);
+		await wmo.exportAsOBJ(exportPath, helper);
+
+		// Abort if the export has been cancelled.
+		if (helper.isCancelled())
+			return;
 
 		helper.mark(fileName, true);
 	} catch (e) {
@@ -213,6 +216,10 @@ const exportSelectedMap = async () => {
 	const markPath = path.join('maps', selectedMapDir, selectedMapDir);
 
 	for (const index of exportTiles) {
+		// Abort if the export has been cancelled.
+		if (helper.isCancelled())
+			break;
+
 		const adt = new ADTExporter(selectedMapID, selectedMapDir, index);
 
 		// Locate game objects within the tile for exporting.
@@ -230,7 +237,7 @@ const exportSelectedMap = async () => {
 		}
 
 		try {
-			await adt.export(dir, exportQuality, gameObjects);
+			await adt.export(dir, exportQuality, gameObjects, helper);
 			helper.mark(markPath, true);
 		} catch (e) {
 			helper.mark(markPath, false, e.message);
@@ -267,7 +274,7 @@ core.events.once('screen-tab-maps', async () => {
 	for (const [id, entry] of table.getAllRows()) {
 		const wdtPath = util.format('world/maps/%s/%s.wdt', entry.Directory, entry.Directory);
 		if (listfile.getByFilename(wdtPath))
-			maps.push(util.format('[%d]\x19[%d]\x19%s\x19(%s)', entry.ExpansionID, id, entry.MapName_lang, entry.Directory));
+			maps.push(util.format('%d\x19[%d]\x19%s\x19(%s)', entry.ExpansionID, id, entry.MapName_lang, entry.Directory));
 	}
 
 	core.view.mapViewerMaps = maps;