+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Update three.js.
 
 0.1.20 (In Development)
+- Added expansion icons to the map exporter list.
+- Added ability to cancel exports that are in-progress.
 - Fixed an issue exporting pre-baked map tiles that lack height textures. (example: https://i.imgur.com/v9nRgjk.jpg)
 - Fixed an issue that caused the toast bar to disappear while an export was in progress.
 - Using the automatic Blender add-on installer now targets all installed versions of Blender on your system.